Main screen module, with added signup screen
