--- GoogleService-Info.plist
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>API_KEY</key>
-	<string>AIzaSyAubbmFYEA3K76ACclql2t6UG9gQLPe0yU</string>
-	<key>GCM_SENDER_ID</key>
-	<string>712205756102</string>
-	<key>PLIST_VERSION</key>
-	<string>1</string>
-	<key>BUNDLE_ID</key>
-	<string>com.mossyskies.app</string>
-	<key>PROJECT_ID</key>
-	<string>mossy-skies</string>
-	<key>STORAGE_BUCKET</key>
-	<string>mossy-skies.firebasestorage.app</string>
-	<key>IS_ADS_ENABLED</key>
-	<false></false>
-	<key>IS_ANALYTICS_ENABLED</key>
-	<false></false>
-	<key>IS_APPINVITE_ENABLED</key>
-	<true></true>
-	<key>IS_GCM_ENABLED</key>
-	<true></true>
-	<key>IS_SIGNIN_ENABLED</key>
-	<true></true>
-	<key>GOOGLE_APP_ID</key>
-	<string>1:712205756102:ios:09d18a421c88f58d2859c7</string>
-</dict>
-</plist>
